requirements-ci.txt
opencv-python
numpy
paho-mqtt
streamlit
onnxruntime
